+++ src/test/java/com/rollingpaper/ggeujeogggeujeog/user/application/UserServiceTest.java
@@ -6,8 +6,6 @@ import static org.mockito.BDDMockito.*;
 
 import java.util.Optional;
 
-import javax.servlet.http.HttpSession;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -19,7 +17,6 @@ import com.rollingpaper.ggeujeogggeujeog.common.util.PasswordEncoder;
 import com.rollingpaper.ggeujeogggeujeog.user.exception.DuplicatedEmailException;
 import com.rollingpaper.ggeujeogggeujeog.user.exception.NoSuchUserException;
 import com.rollingpaper.ggeujeogggeujeog.user.infrastructure.UserMapper;
-import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.SignInRequestDto;
 import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.SignUpRequestDto;
 import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.UserUpdateRequestDto;
 
@@ -34,9 +31,6 @@ class UserServiceTest {
 
 	@Mock
 	private PasswordEncoder passwordEncoder;
-
-	@Mock
-	private HttpSession httpSession;
 
 	@Test
 	@DisplayName("중복된 이메일이 없다면 회원가입에 성공한다")
@@ -71,43 +65,6 @@ class UserServiceTest {
 		assertThrows(DuplicatedEmailException.class, () -> userService.register(dto));
 	}
 
-	@Test
-	@DisplayName("회원이 존재하고 패스워드가 일치하면 로그인에 성공한다")
-	void signIn() {
-		//given
-		SignInRequestDto dto = new SignInRequestDto(TestUser.USER1.getEmail(), TestUser.USER1.getPassword());
-		given(userMapper.findByEmail(any())).willReturn(Optional.ofNullable(TestUser.USER1));
-		given(passwordEncoder.matches(TestUser.USER1.getPassword(), TestUser.USER1.getPassword())).willReturn(true);
-
-		//when
-		userService.signIn(dto);
-
-		//then
-		then(httpSession).should(times(1)).setAttribute(any(), any());
-	}
-
-	@Test
-	@DisplayName("이메일이 일치하지 않으면 로그인에 실패한다")
-	void signInWithEmailNotFound() {
-		//given
-		SignInRequestDto dto = new SignInRequestDto(TestUser.USER1.getEmail(), TestUser.USER1.getPassword());
-		given(userMapper.findByEmail(any())).willReturn(Optional.ofNullable(null));
-
-		//then
-		assertThrows(NoSuchUserException.class, () -> userService.signIn(dto));
-	}
-
-	@Test
-	@DisplayName("회원은 존재하고 패스워드가 틀릴 경우 로그인에 실패한다")
-	void signInWithPasswordNotFound() {
-		//given
-		SignInRequestDto dto = new SignInRequestDto(TestUser.USER1.getEmail(), TestUser.USER1.getPassword());
-		given(userMapper.findByEmail(any())).willReturn(Optional.ofNullable(TestUser.USER1));
-		given(passwordEncoder.matches(any(), any())).willReturn(false);
-
-		//then
-		assertThrows(IllegalArgumentException.class, () -> userService.signIn(dto));
-	}
 
 	@Test
 	@DisplayName("현재 회원의 정보를 수정한다")
